Cleaned the code to remove warnings.

+++ BaconBox/Input/GamePad/GamePad.cpp
@@ -4,16 +4,16 @@
 
 using namespace BaconBox;
 
-GamePad* GamePad::getDefault() {
+GamePad *GamePad::getDefault() {
 	return InputManager::getDefaultGamePad();
 }
 
-const GamePadState& GamePad::getState() const {
+const GamePadState &GamePad::getState() const {
 	return state;
 }
 
-GamePad::GamePad(int index) : InputDevice() {
-    gamePadIndex = index;   
+GamePad::GamePad(size_t index) : InputDevice() {
+	gamePadIndex = index;
 }
 
 GamePad::~GamePad() {
@@ -24,32 +24,32 @@ void GamePad::updateDevice() {
 }
 
 
-unsigned int GamePad::getIndex(){
-    return gamePadIndex;
+size_t GamePad::getIndex() {
+	return gamePadIndex;
 }
 
 
 
-std::vector<float>& GamePad::getPreviousButtons(){
-    return state.previousButtons;
+std::vector<float> &GamePad::getPreviousButtons() {
+	return state.previousButtons;
 }
 
-std::vector<float>& GamePad::getButtons(){
-    return state.buttons;
+std::vector<float> &GamePad::getButtons() {
+	return state.buttons;
 }
 
-std::vector<float>& GamePad::getThumbstick(){
-    return state.thumbsticks;
+std::vector<float> &GamePad::getThumbstick() {
+	return state.thumbsticks;
 }
 
-std::vector<float>& GamePad::getPreviousThumbstick(){
-    return state.previousThumbsticks;
+std::vector<float> &GamePad::getPreviousThumbstick() {
+	return state.previousThumbsticks;
 }
 
 
 
 bool GamePad::isButtonPressed(unsigned int buttonIndex) const {
-    return state.isButtonPressed(buttonIndex);
+	return state.isButtonPressed(buttonIndex);
 }
 
 bool GamePad::isButtonHeld(unsigned int buttonIndex) const {
@@ -60,10 +60,10 @@ bool GamePad::isButtonReleased(unsigned int buttonIndex) const {
 	return state.isButtonReleased(buttonIndex);
 }
 
-unsigned int GamePad::getNbOfButton(){
-    return getButtons().size();
+size_t GamePad::getNbOfButton() {
+	return getButtons().size();
 }
 
-unsigned int GamePad::getNbOfThumbstick(){
-    return getThumbstick().size();
+size_t GamePad::getNbOfThumbstick() {
+	return getThumbstick().size();
 }
